informative_drawings/requirements.txt
torch
torchvision
torchaudio
numpy
scipy
tensorboard
pillow
opencv-python
scikit-learn
oyaml
dominate
ninja
ftfy
tqdm
regex
